Added poinTime Class, further logic work in Plugin file
poinTime is a data type responsible for storing LocalPoints and the GameTick of said LocalPoint.  each instance of pointTime will be stored in a First In First Out queue with a maximum of 100 datapoints. This is because each game tick in OSRS is approximately 600ms, translating 100 ticks to 60 seconds, the time it takes for an item to appear on the ground.